Update README.md
Let’s improve README for better accessibility 
 
Moved the Contributing Section: Placed it at the top so that individuals new to the page get to see how to get in easily. 
Simplified Content: Organized the Contributing steps and made it as simple and concise as possible with directions for developers and non-coder folk. 
Enhanced Readability: Broke key information down into categories to make the README easier to find specific information. 
These changes are used in order to ensure that contributing to the project is easy and popular among all members.